SilverTier/Silver_Tier_FTE/requirements.txt
secure-smtplib
pywhatkit
python-dotenv
watchdog
openai-agents